Working prototype
Implemented sending macro, domain and aircraft geometry, run shell script and execute using screen to detach the process from the ssh session.
The program then connects to the detached process and if the ssh connection drops, the simulation is unnafected.
Need to use system("cls") to clear the screen before reattaching, for proper console output.